# docs/04-API/openapi.yaml
openapi: 3.0.3
info:
  title: 教务管理系统 API
  description: |
    培训机构**纯员工端 CRM** 后端 API（Express 4 + node:sqlite）。
    仅 `admin` / `teacher` 两种员工角色；学生、家长无账号、不登录。

    本文件是机器可读规范，人类可读的完整说明见 `docs/04-API/API.md`。
    覆盖范围：认证 + 四个 P0 模块（考勤签到 / 学生 / 成绩 / 招生线索）+ 通知 + AI 分析端点 + 课次实体与任课关系（v21）。
    统一响应包裹：成功 `{ success: true, data }`；失败 `{ success: false, message }`。
    `/api/analytics/*` 使用独立的「AI 分析信封」。
    v21（2026-09-23）：新增 `/api/sessions/*`（课次）、`/api/teaching-assignments`（任课关系）、`/api/period-times`（节次时间）；
    `POST /api/attendance/batch`、`GET /api/attendance`、`POST /api/growth/class-eval` 增加可选 `session_id`。
  version: 1.0.0
  contact:
    name: 教务管理系统
servers:
  - url: http://localhost:3000
    description: 本地 / Docker 后端
...
security:
  - bearerAuth: []
paths:
...
  /api/growth/eval-form:
    get:
      tags: [Growth]
      summary: 课后采集预填数据（名单 + 可评知识点 + 今日已评）
      description: |
        老师打开「授课流程」页时的预填数据 —— 不让老师自己找该评什么。

        **`knowledge_points[]` 只含叶子知识点（可评定对象）**，单元节点被后端过滤。
        `knowledge_points` 是两级结构：单元（`seq=0` / `parent_id IS NULL`）与其下的
        知识点（`seq>0` / `parent_id=单元 id`）。单元是分组标题，老师无法打勾；
        不过滤会有两个后果：① 出现无法操作的行；② 单元节点占掉 LIMIT 名额，
        把真正可评的知识点挤出去。单元名以 `unit_name` 带下来做分组标题。
      security:
        - bearerAuth: []
      parameters:
        - name: class_id
          in: query
          required: true
          schema: { type: integer }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date, example: "2026-09-20" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/GrowthEvalForm" }
        "400":
          description: 参数不合法
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "403":
          description: 无权操作该班
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/growth/class-eval:
    post:
      tags: [Growth]
      summary: 提交课堂评价（老师课后 10 秒动作）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClassEvalInput" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      saved: { type: integer, description: 写入/覆盖的评价条数 }
                      appended: { type: integer, description: 向时间轴追加的事件数 }
                      eval_date: { type: string }
                      class_id: { type: integer }
        "403":
          description: 无权操作该班
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/growth/kp-assessment:
    post:
      tags: [Growth]
      summary: 批量知识点评定（老师课后 1 分钟动作）
      description: |
        每条评定都会向时间轴追加一条 `kp_assessment` 事件（含 `from` 上次等级与 `changed`）。
        **即使等级未变也写** —— 这条记录本身就是"老师每节课都看了学生状态"的过程证据，
        跳过未变化的评定会让成长路径退化成"起点→终点"两点，丢失中间过程。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KpAssessmentInput" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      students: { type: integer }
                      records: { type: integer }
        "403":
          description: 无权操作该班
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/growth/students/{id}/timeline:
    get:
      tags: [Growth]
      summary: 单学员成长时间轴
      description: 既有三表（考勤/成绩/课时）UNION + 新事件，已按 `(type, date)` 去重。
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/StudentTimeline" }
        "403":
          description: 非本班学员
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/growth/students/{id}/growth:
    get:
      tags: [Growth]
      summary: 单学员成长画像
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/StudentGrowth" }
        "403":
          description: 非本班学员
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/growth/classes/{id}/growth:
    get:
      tags: [Growth]
      summary: 整班成长概览
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/ClassGrowth" }
        "403":
          description: 无权查看该班
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/growth/knowledge-points:
    get:
      tags: [Growth]
      summary: 知识点列表
      description: |
        返回**含单元节点的完整层级**（带 `parent_id`），供展示层做树/分组用。
        与 `eval-form` 的区别：`eval-form` 返回可评集合（仅叶子），本端点返回完整层级 ——
        两者定位不同，不应合并。
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/KnowledgePoint" }
  /api/growth/thresholds:
    get:
      tags: [Growth]
      summary: 成长阈值（8 条）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/GrowthThresholds" }
    put:
      tags: [Growth]
      summary: 修改成长阈值（仅 admin）
      description: |
        ★ 管理动作 —— 改的是全局阈值，属配置而非教学行为。
        因此这是 `/api/growth` 内**唯一**不对工作台凭证（type=ai_agent）放行的路径，
        工作台调用会得到 403，应改用教务系统管理员账号操作。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: { type: number }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/GrowthThresholds" }
        "403":
          description: 非 admin，或使用了工作台凭证
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 该学员已有报班记录，不能删除 }
...
    KnowledgePoint:
      type: object
      description: |
        知识点（两级结构）。`parent_id IS NULL` 表示这是**单元节点**（分组标题，不可评定）；
        非空表示**叶子知识点**（可评定对象）。
      properties:
        id: { type: integer }
        parent_id: { type: integer, nullable: true }
        code: { type: string, example: "U1-2" }
        name: { type: string, example: "SSS 判定" }
        course_id: { type: integer, nullable: true }
        grade: { type: string, nullable: true }
        unit_no: { type: integer, example: 1 }
        seq: { type: integer, description: 单元节点为 0，叶子为 1..n }
        difficulty: { type: integer, minimum: 1, maximum: 5 }
    GrowthEvalForm:
      type: object
      description: 课后采集预填数据
      properties:
        class_id: { type: integer }
        date: { type: string, format: date }
        students:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              no: { type: string, description: 学号 }
              name: { type: string }
              evaluated: { type: boolean, description: 今日是否已评 }
              eval:
                nullable: true
                type: object
                properties:
                  focus: { type: integer, minimum: 1, maximum: 5 }
                  participation: { type: integer, minimum: 1, maximum: 5 }
                  mastery: { type: integer, minimum: 1, maximum: 5 }
                  note: { type: string }
        knowledge_points:
          type: array
          description: ★ 只含叶子知识点（可评定对象），单元节点已过滤；最多 40 条
          items:
            type: object
            properties:
              id: { type: integer }
              code: { type: string }
              name: { type: string }
              unit_no: { type: integer }
              unit_name: { type: string, nullable: true, description: 所属单元名，供分组标题用 }
              seq: { type: integer }
              difficulty: { type: integer }
        already_evaluated: { type: integer }
    ClassEvalInput:
      type: object
      required: [class_id, eval_date, items]
      properties:
        class_id: { type: integer }
        course_id: { type: integer, nullable: true }
        eval_date: { type: string, format: date }
        session_no: { type: integer, nullable: true }
        items:
          type: array
          items:
            type: object
            required: [student_id]
            properties:
              student_id: { type: integer }
              focus: { type: integer, minimum: 1, maximum: 5, description: 越界自动夹取，缺省 3 }
              participation: { type: integer, minimum: 1, maximum: 5 }
              mastery: { type: integer, minimum: 1, maximum: 5 }
              note: { type: string }
    KpAssessmentInput:
      type: object
      required: [class_id, assessed_at, kps]
      properties:
        class_id: { type: integer }
        course_id: { type: integer, nullable: true }
        assessed_at: { type: string, format: date }
        session_no: { type: integer, nullable: true }
        kps:
          type: array
          items:
            type: object
            required: [kp_id, level]
            properties:
              kp_id: { type: integer }
              level:
                type: string
                enum: [未掌握, 部分掌握, 已掌握]
        student_ids:
          type: array
          description: 缺省 = 全班在读学员
          items: { type: integer }
        overrides:
          type: array
          description: 个别学员单独调整
          items:
            type: object
            properties:
              student_id: { type: integer }
              kp_id: { type: integer }
              level:
                type: string
                enum: [未掌握, 部分掌握, 已掌握]
    StudentTimeline:
      type: object
      properties:
        studentId: { type: integer }
        from: { type: string }
        to: { type: string }
        count: { type: integer }
        events:
          type: array
          items:
            type: object
            properties:
              type: { type: string, example: "kp_assessment" }
              label: { type: string }
              date: { type: string }
              summary: { type: string }
              payload: { type: object, additionalProperties: true }
    StudentGrowth:
      type: object
      description: 单学员成长画像
      properties:
        studentId: { type: integer }
        attendance: { type: object, additionalProperties: true }
        evalTrend: { type: array, items: { type: object, additionalProperties: true } }
        examSeries: { type: array, items: { type: object, additionalProperties: true } }
        kpProgress: { type: array, items: { type: object, additionalProperties: true } }
        milestones: { type: array, items: { type: object, additionalProperties: true } }
        range:
          type: object
          properties:
            from: { type: string, description: 无数据时为哨兵值 0000-01-01，前端需自行回落 }
            to: { type: string, description: 无数据时为哨兵值 9999-12-31，前端需自行回落 }
    ClassGrowth:
      type: object
      description: 整班成长概览
      properties:
        classId: { type: integer }
        studentCount: { type: integer }
        students:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              student_no: { type: string }
              name: { type: string }
        evalSeries:
          type: array
          items: { type: object, additionalProperties: true }
        kpMastery:
          type: array
          items: { type: object, additionalProperties: true }
        thresholds:
          type: object
          additionalProperties: { type: number }
    GrowthThresholds:
      type: object
      description: 成长阈值（8 条默认值，admin 可调，不重建镜像）
      additionalProperties: { type: number }
...